Move settings endpoints to their own module
